Conflicts:
	dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/refbook/filter/components/ForeignKeyResolverComponent.java

+++ gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/audit/server/AuditArchiveHandler.java
@@ -1,11 +1,9 @@
 package com.aplana.sbrf.taxaccounting.web.module.audit.server;
 
-import com.aplana.sbrf.taxaccounting.model.LogSearchResultItem;
-import com.aplana.sbrf.taxaccounting.model.PagingResult;
+import com.aplana.sbrf.taxaccounting.model.*;
 import com.aplana.sbrf.taxaccounting.model.exception.ServiceException;
-import com.aplana.sbrf.taxaccounting.service.AuditService;
-import com.aplana.sbrf.taxaccounting.service.BlobDataService;
-import com.aplana.sbrf.taxaccounting.service.PrintingService;
+import com.aplana.sbrf.taxaccounting.model.log.Logger;
+import com.aplana.sbrf.taxaccounting.service.*;
 import com.aplana.sbrf.taxaccounting.web.main.api.server.SecurityService;
 import com.aplana.sbrf.taxaccounting.web.module.audit.shared.AuditArchiveAction;
 import com.aplana.sbrf.taxaccounting.web.module.audit.shared.AuditArchiveResult;
@@ -19,6 +17,7 @@ import org.springframework.stereotype.Service;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.util.Date;
 
 /**
  * User: avanteev
@@ -39,28 +38,49 @@ public class AuditArchiveHandler extends AbstractActionHandler<AuditArchiveActio
     @Autowired
     SecurityService securityService;
 
+    @Autowired
+    TAUserService taUserService;
+
+    @Autowired
+    private LogEntryService logEntryService;
+
     public AuditArchiveHandler() {
         super(AuditArchiveAction.class);
     }
 
     @Override
     public AuditArchiveResult execute(AuditArchiveAction action, ExecutionContext context) throws ActionException {
+        TAUserInfo userInfo = securityService.currentUserInfo();
         AuditArchiveResult result = new AuditArchiveResult();
-        PagingResult<LogSearchResultItem> records = auditService.getLogsByFilter(action.getLogSystemFilter());
-        if (records.isEmpty())
-            throw new ServiceException("Нет записей за указанную дату.");
-        File filePath = new File(printingService.generateAuditCsv(records));
-        try {
-            String uuid = blobDataService.createTemporary(new FileInputStream(filePath), filePath.getName());
-            result.setUuid(uuid);
-            auditService.removeRecords(records, securityService.currentUserInfo());
-            result.setCountOfRemoveRecords(records.getTotalCount());
-            return result;
-        } catch (FileNotFoundException e) {
-            throw new ServiceException("Возникла проблема при считывании файла, файл не найден.");
-        }finally {
-            filePath.delete();
+        LockData lockData = auditService.lock(userInfo);
+        Logger logger = new Logger();
+        if (lockData == null) {
+            try {
+                PagingResult<LogSearchResultItem> records = auditService.getLogsByFilter(action.getLogSystemFilter());
+                if (records.isEmpty())
+                    throw new ServiceException("Нет записей за указанную дату.");
+                File filePath = new File(printingService.generateAuditCsv(records));
+                try {
+                    String uuid = blobDataService.createTemporary(new FileInputStream(filePath), filePath.getName());
+                    result.setFileUuid(uuid);
+                    auditService.removeRecords(records, securityService.currentUserInfo());
+                    result.setCountOfRemoveRecords(records.getTotalCount());
+                    return result;
+                } catch (FileNotFoundException e) {
+                    throw new ServiceException("Возникла проблема при считывании файла, файл не найден.");
+                } finally {
+                    filePath.delete();
+                }
+            } finally{
+                auditService.unlock(userInfo);
+            }
+        } else {
+            TAUser user = taUserService.getUser((int) lockData.getUserId());
+            logger.error("Операция недоступна, так как она выполняется сейчас пользователем " + user.getName() + ". Повторите архивацию позже.");
+            result.setException(true);
+            result.setUuid(logEntryService.save(logger.getEntries()));
         }
+        return result;
     }
 
     @Override
